iommu/of: Call pci_request_acs() before enumerating the Root Port device
When booting with devicetree, ACS is enabled for all the ACS capable PCI
devices other than the first Root Port enumerated in the system. This is
due to calling pci_request_acs() after the enumeration and initialization
of the Root Port device. This prevents ACS from being enabled for the Root
Port device. But afterwards, ACS is getting enabled for the rest of the PCI
devices since pci_request_acs() sets the 'pci_acs_enable' flag and the PCI
core uses this flag to enable ACS for all ACS capable devices.

Ideally, pci_request_acs() should only be called if the 'iommu-map' DT
property is set for the host bridge device. Hence, call pci_request_acs()
from devm_of_pci_bridge_init() if the 'iommu-map' property is present in
the host bridge DT node. This aligns with the implementation of the ARM64
ACPI driver (drivers/acpi/arm64/iort.c) as well.

With this change, ACS will be enabled for all the PCI devices including the
Root Port device.

Fixes: 6bf6c24720d33 ("iommu/of: Request ACS from the PCI core when configuring IOMMU linkage")
Signed-off-by: Xingang Wang <wangxingang5@huawei.com>
Signed-off-by: Pavankumar Kondeti <quic_pkondeti@quicinc.com>
Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@oss.qualcomm.com>
Link: https://patch.msgid.link/20241107-pci_acs_fix-v1-1-185a2462a571@quicinc.com