dt-bindings: net: wireless: ath12k: Make PMU supplies optional for M.2 connector
These PMU supplies are only needed if Qcom proprietary connector is used to
connect the ath12k devices like WCN7850 to the PCIe host. If M.2 connector
is used, these PMU supplies are not needed. The M.2 connector spec defines
the standard supplies (VDD 3.3v and VIO 1.8v) for the slot, and the device
will internally derive the PMU supplies from the 3.3v supply. These
standard supplies will be part of the PCI slot/bus node, so the device node
will not have any supplies.

So make the PMU supplies optional for using the device with M.2 connector.

Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@oss.qualcomm.com>