PCI/pwrctrl: slot: Add support for the optional W_DISABLE1# signal
PCIe M.2 spec r4.0 defines the W_DISABLE1# signal for Socket 1 and Socket 2
form factors. If present, this signal will be used to control the radio on
the M.2 card for WWAN or WLAN solutions. Hence, add support for controlling
this signal if the platform uses GPIO for it. This signal will be driven
low while powering on the slot and driven high while powering it off.

Note: There is one more W_DISABLE2# signal defined by the M.2 spec, but it
is mostly used for Bluetooth and GNSS interfaces, not related to PCI. So it
is not supported by this driver.

Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@oss.qualcomm.com>