PCI/pwrctrl: Add helpers for creating and destroying pwrctrl devices
So far, the PCI core created the pwrctrl devices while scanning for a
device in pci_scan_device(). Once the device is created, PCI core will skip
enumerating the device in a hope that the pwrctrl driver will power on the
device and it will trigger rescan of the bus.

This design worked well for endpoint devices that were directly connected
to the Root Ports. But when a PCIe bus extender like PCIe Switch is
connected to the Root Port and if it needs the pwrctrl driver support, this
design doesn't fit well. The reason is, once the PCI enumeration for the
switch is skipped, the bus scan will proceed further and the PCI core will
eventually complete the scan before the switch device gets powered on and
enumerated. Once the bus can has completed, the PCI core would've allocated
the resources like bridge window and bus numbers to the upstream bridge
(Root Port for instance). And the allocated resource might not be
sufficient if the upstream bridge is not hotplug capable. The PCI core will
only allocate the resource for bridge based on the number of downstream
busses available at the time of bus scan. If the bridge is not hotplug
capable, it will most likely allocate resource for one downstream bus under
the bridge. This resource will work fine for an endpoint, but for switch,
it will not be sufficient since a switch may have multiple downstream
ports.

So currently, a PCIe switch cannot be enumerated properly if it is powered
on by the pwrctrl driver in kernel. The vendors have to power on the switch
in the bootloader to workaround this issue. But it defeats the purpose of
using the pwrctrl drivers.

This issue is fixed in two parts. First this commit adds a new set of
helpers that create and destroy the pwrctrl devices when called. These
helpers will recursively scan the PCI child nodes of the controller node
and create the pwrctrl device if required (based on the same criteria used
earlier).

The second part will be handled in the next commit.

Signed-off-by: Krishna Chaitanya Chundru <krishna.chundru@oss.qualcomm.com>
[mani: splitted the patch, cleaned up the code, and rewrote description]
Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@oss.qualcomm.com>